Added DownloadImage.tsx: restruct function ofuct function of download
